Moved two states and two functions used in 5 components to the custom useModal hook.
