+++ src/helpers/fetchingFunctions.ts
